refactor(transform): move files to dedicated folders
