+++ docker-compose.override.yml.template
@@ -9,7 +9,7 @@ services:
         ports:
             - "127.0.0.1:8080:80"
         # Uncomment for dev (hot reload):
-        # command: [ "fastapi", "dev", "--host", "0.0.0.0", "transform.py", "--port", "80" ]
+        # command: [ "fastapi", "dev", "--host", "0.0.0.0", "app/routes.py", "--port", "80" ]
 
     celery:
         # Uncomment for dev (auto-restart on file changes):
@@ -19,7 +19,7 @@ services:
         #     "--directory=/code/src",
         #     "--directory=/code/config",
         #     "--pattern=*.py", "--recursive", "--",
-        #     "celery", "-A", "tasks", "worker",
+        #     "celery", "-A", "transform.tasks", "worker",
         #     "-E", "--pool=solo", "--loglevel=DEBUG"
         #   ]
