+++ docker-compose.yml
@@ -61,7 +61,7 @@ services:
           [
               "celery",
               "-A",
-              "tasks",
+              "transform.tasks",
               "worker",
             #"-Q",
             #"batches",
@@ -97,7 +97,7 @@ services:
             - ./src:/code/src
             - ./config:/code/config
             - ./cache/fastembed:/root/.cache/fastembed
-        command: [ "fastapi", "run", "--host", "0.0.0.0", "transform.py", "--port", "80" ]
+        command: [ "fastapi", "run", "--host", "0.0.0.0", "app/routes.py", "--port", "80" ]
         networks:
             - warehouse-backend
         environment:
